Fold C-PRISM into the SMatrix API as upsample= and tolerance= options
Review feedback on this branch asked that C-PRISM not be a separate pipeline
with its own entry points (SMatrix.scan vs CPRISM.scan), but be integrated
into the existing API with keywords. The CPRISM builder class is removed;
the algorithm is now enabled directly on SMatrix:

    # PRISM (unchanged)
    SMatrix(potential=p, energy=200e3, semiangle_cutoff=20, interpolation=4)

    # C-PRISM: same class, same build/scan/reduce - two keywords
    SMatrix(potential=p, energy=200e3, semiangle_cutoff=20, interpolation=4,
            upsample=True, tolerance=1e-3)

- SMatrix gains upsample, tolerance, max_rank, window_gpts and
  position_quantization keywords; with upsample=False (the default) every
  code path and result is unchanged.
- SMatrix.build returns a CompressedSMatrixArray (the renamed CPRISMArray)
  when upsample=True, and a plain SMatrixArray otherwise or at an
  interpolation factor of (1, 1), where the expansion is already complete.
  Both expose the same scan/reduce interface through BaseSMatrix.
- SMatrix.reduce/scan keep their signatures; when upsampling they force the
  disable_s_matrix_chunks path (the compression requires the scattering
  matrix in memory), reusing the existing eager/lazy per-configuration
  orchestration instead of the duplicated copies CPRISM carried. Frozen
  phonons, ctf ensembles, and lazy evaluation work as before.
- The wave-vector, downsampling and window geometry branch on the option,
  and the compression pipeline moves into abtem/prism/s_matrix.py; the
  separate abtem/prism/c_prism.py module is deleted.
- Tests move to test/test_prism_upsample.py under the new API, plus new
  tests of the option validation and the (1, 1) PRISM identity.

Verified bit-for-bit against the previous CPRISM implementation on scans at
interpolation 1 and 2 (eager and lazy, frozen phonons, windowed reduction,
single off-grid positions). Full test suite: 832 passed.
