+++ abtem/__init__.py
@@ -24,8 +24,7 @@ from abtem.measurements import (
     ReciprocalSpaceLineProfiles,
 )
 from abtem.potentials.iam import CrystalPotential, Potential, PotentialArray
-from abtem.prism.c_prism import CPRISM, CPRISMArray
-from abtem.prism.s_matrix import SMatrix, SMatrixArray
+from abtem.prism.s_matrix import CompressedSMatrixArray, SMatrix, SMatrixArray
 from abtem.scan import CustomScan, GridScan, LineScan
 from abtem.transfer import CTF, Aperture, SpatialEnvelope, TemporalEnvelope
 from abtem.visualize.visualizations import show_atoms
@@ -53,10 +52,9 @@ __all__ = [
     "MeasurementsEnsemble",
     "PolarMeasurements",
     "IndexedDiffractionPatterns",
-    "CPRISM",
-    "CPRISMArray",
     "SMatrix",
     "SMatrixArray",
+    "CompressedSMatrixArray",
     "FrozenPhonons",
     "AtomsEnsemble",
     "Potential",
